fix: type of id field in base model

+++ models/base.go
@@ -8,7 +8,7 @@ import (
 Abstract Model
 */
 type Base struct {
-	ID        int64     `json:"id" gorm:"primaryKey"`
+	ID        int       `json:"id" gorm:"primaryKey"`
 	CreatedAt time.Time `json:"created_at"`
 	UpdatedAt time.Time `json:"updated_at"`
 }
